reduce matrix/array copying
I’ve not tested these things, and did the changes without examining the
code in any depth, so it’s possible errors will crop up.

+++ R/adp.R
@@ -1870,7 +1870,7 @@ binmapAdp <- function(x, debug=getOption("oceDebug"))
     ## vectorizing across the loop, by combining into a single vector
     ## for (distance, cr, ...), but it was no faster, and the code was
     ## more complicated to read.
-    vbm <- array(dim=dim(v))
+    vbm <- array(double(), dim=dim(v))
     abm <- array(raw(), dim=dim(v))
     qbm <- array(raw(), dim=dim(v))
     gbm <- array(raw(), dim=dim(v))
